+++ Gui/RightClickMenu/AchievementMenu/FirstCategory.lua
@@ -1,5 +1,4 @@
 local _, addon = ...;
-local menuUtil = addon.Gui.MenuUtil;
 local section = {};
 tinsert(addon.Gui.RightClickMenu.AchievementMenu:GetLastSection().Sections, section);
 
@@ -9,6 +8,6 @@ function section:CheckAdd()
 end
 
 function section:Add(menu, achievement)
-	menuUtil:CreateTitle(menu, achievement.Category:GetPath());
+	addon.MenuUtil:CreateTitle(menu, achievement.Category:GetPath());
 	addon.Gui.RightClickMenu.AchievementMenu:AddGoToAchievementLine(menu, achievement.Id);
 end