+++ Gui/RightClickMenu/AchievementMenu/GoTo.lua
@@ -1,5 +1,4 @@
 local _, addon = ...;
-local menuUtil = addon.Gui.MenuUtil;
 local section = {
 	Sections = {}
 };
@@ -15,18 +14,18 @@ function section:CheckAdd(achievement)
 end
 
 function section:Add(menu, achievement)
-	local childMenu = menuUtil:CreateButton(menu, addon.L["Go to"]);
+	local childMenu = addon.MenuUtil:CreateButton(menu, addon.L["Go to"]);
 	local addSeparator = nil;
 
 	for _, sect in next, self.Sections do
 		if sect:CheckAdd(achievement) then
 			if addSeparator then
-				menuUtil:CreateDivider(childMenu);
+				addon.MenuUtil:CreateDivider(childMenu);
 			end
 			sect:Add(childMenu, achievement);
 			addSeparator = true;
 		end
 	end
 
-    menuUtil:AddChildMenu(menu, childMenu);
+    addon.MenuUtil:AddChildMenu(menu, childMenu);
 end