+++ Gui/RightClickMenu/AchievementMenu/OtherFaction.lua
@@ -1,5 +1,4 @@
 local _, addon = ...;
-local menuUtil = addon.Gui.MenuUtil;
 local section = {};
 tinsert(addon.Gui.RightClickMenu.AchievementMenu:GetLastSection().Sections, section);
 
@@ -14,6 +13,6 @@ function section:Add(menu, achievement)
 	elseif addon.Faction.IsHorde then
 		faction = addon.L["Alliance"];
 	end
-	menuUtil:CreateTitle(menu, addon.L["Other faction"] .. " (" .. faction .. ")");
+	addon.MenuUtil:CreateTitle(menu, addon.L["Other faction"] .. " (" .. faction .. ")");
 	addon.Gui.RightClickMenu.AchievementMenu:AddGoToAchievementLine(menu, achievement.OtherFactionAchievementId);
 end