+++ Gui/RightClickMenu/AchievementMenu/OtherLocations.lua
@@ -1,5 +1,4 @@
 local _, addon = ...;
-local menuUtil = addon.Gui.MenuUtil;
 local section = {};
 tinsert(addon.Gui.RightClickMenu.AchievementMenu:GetLastSection().Sections, section);
 
@@ -8,7 +7,7 @@ function section:CheckAdd(achievement)
 end
 
 function section:Add(menu, achievement)
-	menuUtil:CreateTitle(menu, addon.L["Other locations"]);
+	addon.MenuUtil:CreateTitle(menu, addon.L["Other locations"]);
 	local selectedCategory = addon.Gui.SelectedTab and addon.Gui.SelectedTab.SelectedCategory or nil;
 	if selectedCategory ~= achievement.Category then
 		addon.Gui.RightClickMenu.AchievementMenu:AddGoToAchievementWithCategoryLine(menu, achievement, achievement.Category);