+++ Gui/RightClickMenu/AchievementMenu/PartOfAChain.lua
@@ -1,5 +1,4 @@
 local _, addon = ...;
-local menuUtil = addon.Gui.MenuUtil;
 local section = {};
 tinsert(addon.Gui.RightClickMenu.AchievementMenu:GetLastSection().Sections, section);
 
@@ -28,6 +27,6 @@ local function AddPartOfAChainAchievement(menu, id, nameSuffix)
 end
 
 function section:Add(menu)
-	menuUtil:CreateTitle(menu, addon.L["Part of a chain"]);
+	addon.MenuUtil:CreateTitle(menu, addon.L["Part of a chain"]);
     AddPartOfAChainAchievement(menu, firstAchievement.Id);
 end