+++ Gui/RightClickMenu/AchievementMenu/RequiredFor.lua
@@ -1,5 +1,4 @@
 local _, addon = ...;
-local menuUtil = addon.Gui.MenuUtil;
 local section = {};
 tinsert(addon.Gui.RightClickMenu.AchievementMenu:GetLastSection().Sections, section);
 
@@ -10,7 +9,7 @@ function section:CheckAdd(achievement)
 end
 
 function section:Add(menu, achievement)
-	menuUtil:CreateTitle(menu, addon.L["Required for"]);
+	addon.MenuUtil:CreateTitle(menu, addon.L["Required for"]);
 	for _, id in next, requiredForIds do
 		if id ~= achievement.Id then
 			addon.Gui.RightClickMenu.AchievementMenu:AddGoToAchievementLine(menu, id);