+++ Gui/RightClickMenu/AchievementMenu/TransmogSets.lua
@@ -1,5 +1,4 @@
 local _, addon = ...;
-local menuUtil = addon.Gui.MenuUtil;
 local section = {};
 tinsert(addon.Gui.RightClickMenu.AchievementMenu.Sections, section);
 
@@ -8,12 +7,12 @@ function section:CheckAdd(achievement)
 end
 
 function section:Add(menu, achievement)
-	local transmogSets = menuUtil:CreateButton(menu, addon.L["Transmog Sets"]);
+	local transmogSets = addon.MenuUtil:CreateButton(menu, addon.L["Transmog Sets"]);
 
 	local tSets = addon.GetUsableSets(achievement.TransmogSets);
 	for _, set in next, tSets do
 		local setInfo = C_TransmogSets.GetSetInfo(set.Id);
-		menuUtil:CreateButtonAndAdd(
+		addon.MenuUtil:CreateButtonAndAdd(
 			transmogSets,
 			setInfo.name .. " (" .. setInfo.description .. ")",
 			function()
@@ -28,5 +27,5 @@ function section:Add(menu, achievement)
 		);
 	end
 
-    menuUtil:AddChildMenu(menu, transmogSets);
+    addon.MenuUtil:AddChildMenu(menu, transmogSets);
 end