+++ Gui/RightClickMenu/AchievementMenu/WowheadLink.lua
@@ -1,5 +1,4 @@
 local _, addon = ...;
-local menuUtil = addon.Gui.MenuUtil;
 local section = {};
 tinsert(addon.Gui.RightClickMenu.AchievementMenu.Sections, section);
 
@@ -38,7 +37,7 @@ function section:Add(menu, achievement)
 	end
 
 	local externalLink = "https://" .. locale .. "wowhead.com/" .. expansion .. "achievement=" .. achievement.Id .. relatedTab;
-    menuUtil:CreateButtonAndAdd(
+    addon.MenuUtil:CreateButtonAndAdd(
 		menu,
 		addon.L["Wowhead"],
 		function()