+++ Gui/RightClickMenu/EventReminderMenu/EventReminderMenu.lua
@@ -1,12 +1,11 @@
 local _, addon = ...;
-local menuUtil = addon.Gui.MenuUtil;
 addon.Gui.RightClickMenu.EventReminderMenu = {
 	Sections = {}
 };
 local eventReminderMenu = addon.Gui.RightClickMenu.EventReminderMenu;
 
 local function CreateMenu(self, menu, event)
-	menuUtil:CreateTitle(menu, event.EventDetails and event.EventDetails.Name or addon.L["Unknown"]);
+	addon.MenuUtil:CreateTitle(menu, event.EventDetails and event.EventDetails.Name or addon.L["Unknown"]);
 
 	for _, section in next, self.Sections do
 		if section:CheckAdd(event) then